Add spear and fix rasterizer state allocations
A new rasterizer state was being created every time the GraphicsDevice.ClearState extensions was being called. I assumed it was a struct, but it was a class so that ended up creating lots of garbage since this extensions gets called so much. We don't do multi-sampling anymore so it wasn't necessary anyway.